+++ misp/main.py
@@ -35,6 +35,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 from typing import Dict, List
 from datetime import datetime, timedelta
+from netskope.common.utils import add_user_agent
 from netskope.integrations.cte.models import Indicator, TagIn
 from netskope.integrations.cte.plugin_base import (
     PluginBase,
@@ -49,25 +50,146 @@ from netskope.integrations.cte.models.business_rule import (
     Action,
 )
 import requests
+import json
+import os
+import traceback
+
+PLATFORM_NAME = "Misp"
+MODULE_NAME = "CTE"
+PLUGIN_VERSION = "1.2.0"
+BATCH_SIZE = 2500
+
 
 class MISPPlugin(PluginBase):
     """The MISP plugin implementation."""
+    def __init__(
+        self,
+        name,
+        *args,
+        **kwargs,
+    ):
+        """Init function.
+
+        Args:
+            name (str): Configuration Name.
+        """
+        super().__init__(
+            name,
+            *args,
+            **kwargs,
+        )
+        self.plugin_name, self.plugin_version = self._get_plugin_info()
+        self.log_prefix = f"{MODULE_NAME} {self.plugin_name}"
+        if name:
+            self.log_prefix = f"{self.log_prefix} [{name}]"
+        
+    def _add_user_agent(self, headers=None) -> str:
+        """Add Client Name to request plugin make.
+
+        Returns:
+            str: String containing the Client Name.
+        """
+        headers = add_user_agent(headers)
+        ce_added_agent = headers.get("User-Agent", "netskope-ce")
+        user_agent_str = "{}-{}-{}-v{}".format(
+            ce_added_agent,
+            MODULE_NAME.lower(),
+            self.plugin_name.lower(),
+            self.plugin_version,
+        )
+
+        headers["User-Agent"] = user_agent_str
+        return headers
+
+    def _get_plugin_info(self) -> tuple:
+        """Get plugin name and version from manifest.
+
+        Returns:
+            tuple: Tuple of plugin's name and version fetched from manifest.
+        """
+        try:
+            file_path = os.path.join(
+                str(os.path.dirname(os.path.abspath(__file__))),
+                "manifest.json",
+            )
+            with open(file_path, "r") as manifest:
+                manifest_json = json.load(manifest)
+                plugin_name = manifest_json.get("name", PLATFORM_NAME)
+                plugin_version = manifest_json.get("version", PLUGIN_VERSION)
+                return (plugin_name, plugin_version)
+        except Exception as exp:
+            self.logger.info(
+                message=(
+                    f"{MODULE_NAME} {PLATFORM_NAME}: Error occurred while"
+                    " getting plugin details."
+                ),
+                details=str(exp),
+            )
+        return (PLATFORM_NAME, PLUGIN_VERSION)
+    
+    def parse_response(self, response):
+        """Parse Response will return JSON from response object.
+
+        Args:
+            response (response): Response object.
+
+        Returns:
+            json: Response Json.
+        """
+        try:
+            return response.json()
+        except json.JSONDecodeError as err:
+            err_msg = (
+                "Invalid JSON response received. "
+                "Error: {}".format(err)
+            )
+            self.logger.error(f"{self.log_prefix}: {err_msg}")
+            raise err
+        except Exception as exp:
+            err_msg = (
+                "Unexpected error occurred while parsing"
+                " json response. Error: {}".format(exp)
+            )
+            self.logger.error(
+                message=f"{self.log_prefix}: {err_msg}"
+            )
+            raise err
 
     def pull(self) -> List[Indicator]:
         """Pull indicators from MISP."""
-        start_time = self.last_run_at  # datetime.datetime object.
+        existing_tag = f"{self.name} Latest"
+        pulling_mechanism = self.configuration.get("pulling_mechanism", "incremental")
+        look_back = self.configuration.get("look_back")
         end_time = datetime.now()
+        tag_utils = TagUtils()
 
-        if not start_time:
-            self.logger.info(
-                f"MISP Plugin: This is initial data fetch since "
-                f"checkpoint is empty. Querying indicators for last {self.configuration['days']} days."
-            )
-            start_time = datetime.now() - timedelta(
-                days=int(self.configuration["days"])
-            )
+        if pulling_mechanism == "look_back":
+            if not look_back:
+                err_msg = "When 'Pulling Mechanism' configuration parameter is set to 'Look Back', please provide a valid non-zero integer value in the 'Look Back' configuration parameter."
+                self.logger.error(f"{self.log_prefix}: {err_msg}")
+                raise err_msg
+            else:
+                # Removing the <config_name> Latest tag from the existing indicators
+                query = {"sources": {"$elemMatch": {"source": f"{self.name}"}}}
+                TagUtils().on_indicators(query).remove(existing_tag)
+                start_time = end_time - timedelta(hours=look_back)
+                if self.last_run_at and self.last_run_at < start_time:
+                    start_time = self.last_run_at
+                new_indicator_tag = self._create_marking_tag(tag_utils, existing_tag)
+        else:
+            start_time = self.last_run_at  # datetime.datetime object.
+
+            if not start_time:
+                self.logger.info(
+                    f"{self.log_prefix}: This is initial data fetch since "
+                    f"checkpoint is empty. Querying indicators for last {self.configuration['days']} days."
+                )
+                start_time = end_time - timedelta(
+                    days=int(self.configuration["days"])
+                )
+        
+        self.logger.info(f"{self.log_prefix}: Start time for this pull cycle: {start_time}")
         # create set of excluded events for
-
         event_ids = []
         if len(self.configuration["include_event_name"]) != 0:
             for inc_event in (
@@ -99,20 +221,46 @@ class MISPPlugin(PluginBase):
         if len(event_ids) != 0:
             body["eventid"] = event_ids
 
-        indicators, tag_utils, skipped_tags = [], TagUtils(), []
+        indicators, skipped_tags = [], []
+
         while True:
-            response = requests.post(
-                f"{self.configuration['base_url'].strip('/')}/attributes/restSearch",
-                headers=self._add_user_agent(
-                    self._get_header(self.configuration["api_key"])
-                ),
-                json=body,
-                verify=self.ssl_validation,
-                proxies=self.proxy,
-            )
+            try:
+                response = requests.post(
+                    f"{self.configuration['base_url'].strip('/')}/attributes/restSearch",
+                    headers=self._add_user_agent(
+                        self._get_header(self.configuration["api_key"])
+                    ),
+                    json=body,
+                    verify=self.ssl_validation,
+                    proxies=self.proxy,
+                )
+            except requests.exceptions.ProxyError as err:
+                self.logger.error(
+                    "{}: Invalid proxy. Error: {}".format(
+                        self.log_prefix, err
+                    ),
+                    details=traceback.format_exc()
+                )
+                raise err
+            except requests.exceptions.ConnectionError as err:
+                self.logger.error(
+                    "{}: Connection error occurred: {}".format(
+                        self.log_prefix, err
+                    ),
+                    details=traceback.format_exc()
+                )
+                raise err
+            except Exception as err:
+                self.logger.error(
+                    "{}: Error occurred while fetching the attributes from the Misp server: {}".format(
+                        self.log_prefix, err
+                    ),
+                    details=traceback.format_exc()
+                )
+                raise err
             response.raise_for_status()
             if response.status_code == 200:
-                json_res = response.json()
+                json_res = self.parse_response(response)
                 body["page"] += 1
 
                 for attr in json_res.get("response", {}).get("Attribute", []):
@@ -141,12 +289,13 @@ class MISPPlugin(PluginBase):
                             self.configuration,
                         )
                         skipped_tags.extend(skipped)
+                        if pulling_mechanism == "look_back" and look_back:
+                            tags.append(new_indicator_tag)
                         indicators.append(
                             Indicator(
                                 value=attr.get("value"),
                                 type=TYPES.get(attr.get("type")),
                                 firstSeen=attr.get("first_seen", None),
-                                lastSeen=attr.get("last_seen", None),
                                 comments=attr.get("comment"),
                                 tags=tags,
                                 extendedInformation=deep_link,
@@ -160,9 +309,29 @@ class MISPPlugin(PluginBase):
                     break
         if len(skipped_tags) > 0:
             self.logger.warn(
-                f"MISP Plugin: Skipping following tag(s) because they are too long: {', '.join(skipped_tags)}"
+                f"{self.log_prefix}: Skipping following tag(s) because they are too long: {', '.join(skipped_tags)}"
             )
         return indicators
+
+    def _create_marking_tag(self, utils: TagUtils, tag_name: str) -> str:
+        """Create new tag in database if required."""
+        try:
+            if not utils.exists(tag_name):
+                utils.create_tag(
+                    TagIn(
+                        name=tag_name,
+                        color="#ED3347",
+                    )
+                )
+                
+        except ValueError as err:
+            self.logger.error(
+                f"{self.log_prefix}: Error occurred while creating an internal tag. Error: {err}",
+                details=traceback.format_exc(),
+            )
+            raise err
+        else:
+            return tag_name
 
     def _create_tags(
         self, utils: TagUtils, tags: List[dict], configuration: dict
@@ -211,31 +380,76 @@ class MISPPlugin(PluginBase):
                 proxies=self.proxy,
             )
             response.raise_for_status()
+            resp_json = self.parse_response(response)
             if (
                 response.status_code == 200
-                and len(response.json().get("response", [])) > 0
+                and len(resp_json.get("response", [])) > 0
             ):
-                return True, response.json().get("response")[0].get(
+                return True, resp_json.get("response")[0].get(
                     "Event", {}
                 ).get("id", None)
             return False, None
+        except requests.exceptions.ProxyError as err:
+            self.logger.error(
+                "{}: Invalid proxy. Error: {}".format(
+                    self.log_prefix, err
+                ),
+                details=traceback.format_exc()
+            )
+            raise err
+        except requests.exceptions.ConnectionError as err:
+            self.logger.error(
+                "{}: Connection error occurred: {}".format(
+                    self.log_prefix, err
+                ),
+                details=traceback.format_exc()
+            )
+            raise err
         except Exception:
             return False, None
 
     def _create_event(self, payload: dict) -> PushResult:
         """Create a new event on MISP instance with given name/info and attributes."""
-        response = requests.post(
-            f"{self.configuration['base_url'].strip('/')}/events/add",
-            headers=self._add_user_agent(
-                self._get_header(self.configuration["api_key"])
-            ),
-            json=payload,
-            verify=self.ssl_validation,
-            proxies=self.proxy,
-        )
-        response.raise_for_status()
+        try:
+            response = requests.post(
+                f"{self.configuration['base_url'].strip('/')}/events/add",
+                headers=self._add_user_agent(
+                    self._get_header(self.configuration["api_key"])
+                ),
+                json=payload,
+                verify=self.ssl_validation,
+                proxies=self.proxy,
+            )
+        except requests.exceptions.ProxyError as err:
+            self.logger.error(
+                "{}: Invalid proxy. Error: {}".format(
+                    self.log_prefix, err
+                ),
+                details=traceback.format_exc()
+            )
+            raise err
+        except requests.exceptions.ConnectionError as err:
+            self.logger.error(
+                "{}: Connection error occurred: {}".format(
+                    self.log_prefix, err
+                ),
+                details=traceback.format_exc()
+            )
+            raise err
+        except Exception as err:
+            self.logger.error(
+                "{}: Error occurred while fetching the attributes from the Misp server: {}".format(
+                    self.log_prefix, err
+                ),
+                details=traceback.format_exc()
+            )
+            raise err
         if response.status_code == 200:
             return PushResult(message="Push Successful.", success=True)
+        self.logger.error(
+            f"{self.log_prefix}: Skipping the batch. Response code: {response.status_code}",
+            details=response.text
+        )
         return PushResult(
             message=f"Could not push indicators to MISP. HTTP status code {response.status_code}.",
             success=False,
@@ -243,18 +457,47 @@ class MISPPlugin(PluginBase):
 
     def _update_event(self, event_id: str, payload: dict) -> PushResult:
         """Update given event's info and attribute(s)."""
-        response = requests.post(
-            f"{self.configuration['base_url'].strip('/')}/events/edit/{event_id}",
-            headers=self._add_user_agent(
-                self._get_header(self.configuration["api_key"])
-            ),
-            json=payload,
-            verify=self.ssl_validation,
-            proxies=self.proxy,
-        )
-        response.raise_for_status()
+        try:
+            response = requests.post(
+                f"{self.configuration['base_url'].strip('/')}/events/edit/{event_id}",
+                headers=self._add_user_agent(
+                    self._get_header(self.configuration["api_key"])
+                ),
+                json=payload,
+                verify=self.ssl_validation,
+                proxies=self.proxy,
+            )
+        except requests.exceptions.ProxyError as err:
+            self.logger.error(
+                "{}: Invalid proxy. Error: {}".format(
+                    self.log_prefix, err
+                ),
+                details=traceback.format_exc()
+            )
+            raise err
+        except requests.exceptions.ConnectionError as err:
+            self.logger.error(
+                "{}: Connection error occurred: {}".format(
+                    self.log_prefix, err
+                ),
+                details=traceback.format_exc()
+            )
+            raise err
+        except Exception as err:
+            self.logger.error(
+                "{}: Error occurred while fetching the attributes from the Misp server: {}".format(
+                    self.log_prefix, err
+                ),
+                details=traceback.format_exc()
+            )
+            raise err
         if response.status_code == 200:
             return PushResult(message="Push Successful.", success=True)
+
+        self.logger.error(
+            f"{self.log_prefix}: Skipping the batch. Response code: {response.status_code}",
+            details=response.text
+        )
         return PushResult(
             message=f"Could not update the MISP Event with given indicators. HTTP status code {response.status_code}.",
             success=False,
@@ -286,17 +529,27 @@ class MISPPlugin(PluginBase):
         exists, event_id = self._event_exists(
             action_dict.get("event_name"), self.configuration
         )
-        if exists:
-            # Push attributes/indicators to existing event
-            return self._update_event(event_id, {"Attribute": attributes})
-        else:
-            # Create new event with all the attributes
-            return self._create_event(
-                {
-                    "info": action_dict.get("event_name"),
-                    "Attribute": attributes,
-                }
+        for i in range(0, len(attributes), BATCH_SIZE):
+            payload = attributes[i : i + BATCH_SIZE]
+            self.logger.info(
+                f"{self.log_prefix}: Attempting to push batch {(i / BATCH_SIZE) + 1} of size {len(payload)}"
             )
+            if exists:
+                # Push attributes/indicators to existing event
+                _ = self._update_event(event_id, {"Attribute": payload})
+            else:
+                # Create new event with all the attributes
+                _ = self._create_event(
+                    {
+                        "info": action_dict.get("event_name"),
+                        "Attribute": payload,
+                    }
+                )
+                exists, event_id = self._event_exists(
+                    action_dict.get("event_name"), self.configuration
+                )
+   
+        return PushResult(message="Push Successful.", success=True)
 
     def validate(self, configuration: dict) -> ValidationResult:
         """Validate the configuration."""
@@ -306,7 +559,7 @@ class MISPPlugin(PluginBase):
             or type(configuration["base_url"]) != str
         ):
             self.logger.error(
-                "Plugin MISP: Validation error occurred. Error: "
+                f"{self.log_prefix}: Validation error occurred. Error: "
                 "Invalid base URL found in the configuration parameters."
             )
             return ValidationResult(
@@ -319,7 +572,7 @@ class MISPPlugin(PluginBase):
             or type(configuration["api_key"]) != str
         ):
             self.logger.error(
-                "Plugin MISP: Validation error occurred. Error: Invalid API key found in the configuration parameters."
+                f"{self.log_prefix}: Validation error occurred. Error: Invalid API key found in the configuration parameters."
             )
             return ValidationResult(
                 success=False, message="Invalid API key provided."
@@ -329,7 +582,7 @@ class MISPPlugin(PluginBase):
             x in ATTRIBUTE_TYPES for x in configuration["attr_type"]
         ):
             self.logger.error(
-                "Plugin MISP: Validation error occurred. Error: "
+                f"{self.log_prefix}: Validation error occurred. Error: "
                 "Invalid attr_type found in the configuration parameters."
             )
             return ValidationResult(
@@ -340,7 +593,7 @@ class MISPPlugin(PluginBase):
             x in ATTRIBUTE_CATEGORIES for x in configuration["attr_category"]
         ):
             self.logger.error(
-                "Plugin MISP: Validation error occurred. Error: "
+                f"{self.log_prefix}: Validation error occurred. Error: "
                 "Invalid attr_category found in the configuration parameters."
             )
             return ValidationResult(
@@ -349,7 +602,7 @@ class MISPPlugin(PluginBase):
 
         if "tags" not in configuration or type(configuration["tags"]) != str:
             self.logger.error(
-                "Plugin MISP: Validation error occurred. Error: "
+                f"{self.log_prefix}: Validation error occurred. Error: "
                 "tags not valid in the configuration parameters."
             )
             return ValidationResult(
@@ -358,7 +611,7 @@ class MISPPlugin(PluginBase):
 
         if "include_event_name" not in configuration:
             self.logger.error(
-                "Plugin MISP: Validation error occurred. Error: "
+                f"{self.log_prefix}: Validation error occurred. Error: "
                 "include_event_name not found in the configuration parameters."
             )
             return ValidationResult(
@@ -375,7 +628,7 @@ class MISPPlugin(PluginBase):
                 event = event.strip()
                 if not event:
                     self.logger.error(
-                        "Plugin MISP: Validation error occurred."
+                        f"{self.log_prefix}: Validation error occurred."
                         " Error: Invalid Event name found in the configuration parameters."
                     )
                     return ValidationResult(
@@ -384,7 +637,7 @@ class MISPPlugin(PluginBase):
 
                 if event == configuration["event_name"].strip():
                     self.logger.error(
-                        f"Plugin MISP: Validation error occurred. Error: {event} "
+                        f"{self.log_prefix}: Validation error occurred. Error: {event} "
                         f"is present in exclude as well as include."
                     )
 
@@ -392,11 +645,16 @@ class MISPPlugin(PluginBase):
                         success=False,
                         message=f"{event} is present in exclude events as well.",
                     )
-                exist = self._event_exists(event, configuration)[0]
-
+                try:
+                    exist = self._event_exists(event, configuration)[0]
+                except:
+                    return ValidationResult(
+                        success=False,
+                        message=f"Error occurred while fetching the {event} from the MISP Instance.",
+                    )
                 if not exist:
                     self.logger.error(
-                        f"Plugin MISP: Validation error occurred. Error: Provided"
+                        f"{self.log_prefix}: Validation error occurred. Error: Provided"
                         f" Event name: {event}, doesn't exist on MISP Instance."
                     )
                     return ValidationResult(
@@ -408,12 +666,43 @@ class MISPPlugin(PluginBase):
             "enable_tagging"
         ] not in ["yes", "no"]:
             self.logger.error(
-                "Plugin MISP: Validation error occurred. Error: Invalid value for 'Enable Tagging' found. "
+                f"{self.log_prefix}: Validation error occurred. Error: Invalid value for 'Enable Tagging' found. "
                 "Value of Enable Tagging should be 'yes' or 'no'."
             )
             return ValidationResult(
                 success=False,
                 message="Invalid value for 'Enable Tagging' provided. Allowed values are 'yes' or 'no'.",
+            )
+            
+        try:
+            if "pulling_mechanism" not in configuration or configuration[
+                "pulling_mechanism"
+            ] not in ["incremental", "look_back"]:
+                self.logger.error(
+                    f"{self.log_prefix}: Validation error occurred. Error: Invalid value for 'Pulling Mechanism' found. "
+                    "Value of Enable Tagging should be 'look_back' or 'incremental'."
+                )
+                return ValidationResult(
+                    success=False,
+                    message="Invalid value for 'Pulling Mechanism' provided. Allowed values are 'look_back' or 'incremental'.",
+                )
+            elif (
+                configuration.get("pulling_mechanism", "incremental") == "look_back" and
+                (not configuration.get("look_back")
+                or int(configuration.get("look_back")) <= 0)
+            ):
+                err_msg = "When 'Pulling Mechanism' configuration parameter is set to 'Look Back', please provide a valid non-zero integer value in the 'Look Back' configuration parameter."
+                self.logger.error(
+                    f"{self.log_prefix}: {err_msg}"
+                )
+                return ValidationResult(
+                    success=False,
+                    message=err_msg,
+                )
+        except ValueError:
+            return ValidationResult(
+                success=False,
+                message="Invalid Look Back provided.",
             )
 
         try:
@@ -423,7 +712,7 @@ class MISPPlugin(PluginBase):
                 or int(configuration["days"]) <= 0
             ):
                 self.logger.error(
-                    "Plugin MISP: Validation error occurred. Error: Invalid days provided."
+                    f"{self.log_prefix}: Validation error occurred. Error: Invalid days provided."
                 )
                 return ValidationResult(
                     success=False,
@@ -444,14 +733,6 @@ class MISPPlugin(PluginBase):
             "Content-Type": "application/json",
         }
 
-    def _add_user_agent(self, header=None):
-        if header is None:
-            return {"User-Agent": "netskope-cte-1.1.0"}
-        else:
-            if "User-Agent" not in header:
-                header["User-Agent"] = "netskope-cte-1.1.0"
-        return header
-
     def _validate_auth(self, configuration: dict) -> ValidationResult:
         """Validate API key by making REST API call."""
         try:
@@ -464,16 +745,21 @@ class MISPPlugin(PluginBase):
                 proxies=self.proxy,
             )
             response.raise_for_status()
-            if response.status_code == 200:
+            resp_json = self.parse_response(response)
+            if response.status_code == 200 and resp_json.get("version"):
                 return ValidationResult(
                     success=True, message="Validation successful."
+                )
+            else:
+                self.logger.error(
+                    f"{self.log_prefix}: Error occurred while validating the credentials."
                 )
 
         except Exception as ex:
             self.logger.error(
-                "MISP: Could not validate authentication credentials."
+                f"{self.log_prefix}: Could not validate authentication credentials. Exception: {ex}",
+                details=traceback.format_exc(),
             )
-            self.logger.error(repr(ex))
         return ValidationResult(
             success=False,
             message="Error occurred while validating account credentials.",