+++ misp/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.2.0
+## Added
+- Added batching of 2500 in push functionality.
+- Added new configuration parameters for defining the pulling mechanism (Incremental or Look Back) and the look back time in hours.
+
 # 1.1.0
 ## Added
 - Added new configuration parameters.